pyproject.toml
[tool.poetry]
name = "fasjson"
version = "0.0.1"
description = "fasjson makes it possible for applications to talk to the fedora account system."

license = "MIT"

authors = [
  "Fedora Infrastructure <admin@fedoraproject.org>"
]

readme = 'README.md'

repository = "https://github.com/fedora-infra/fasjson"
homepage = "https://github.com/fedora-infra/fasjson"

[tool.poetry.dependencies]
python = "^3.6"
Flask = "^1.1.1"
python-ldap = "^3.2.0"
dnspython = "^1.16.0"
gssapi = "^1.6.2"
typing = "^3.7.4.1"
flask-restx = "^0.2.0"

[tool.poetry.dev-dependencies]
flake8 = "^3.7"
pytest = "^5.3.5"
pytest-mock = "^2.0.0"
pytest-cov = "^2.8"
liccheck = "^0.4.2"
bandit = "^1.6"
black = {version = "^18.3-alpha.0", allow-prereleases = true}

[tool.black]
line-length = 78

[tool.towncrier]
package = "fasjson"
filename = "docs/release_notes.rst"
directory = "news/"
title_format = "v{version}"
issue_format = "{issue}"
template = "news/_template.rst"
underlines = "=^-"
wrap = true
all_bullets = true

  [[tool.towncrier.type]]
  directory = "bic"
  name = "Backwards Incompatible Changes"
  showcontent = true

  [[tool.towncrier.type]]
  directory = "dependency"
  name = "Dependency Changes"
  showcontent = true

  [[tool.towncrier.type]]
  directory = "feature"
  name = "Features"
  showcontent = true

  [[tool.towncrier.type]]
  directory = "bug"
  name = "Bug Fixes"
  showcontent = true

  [[tool.towncrier.type]]
  directory = "dev"
  name = "Development Improvements"
  showcontent = true

  [[tool.towncrier.type]]
  directory = "docs"
  name = "Documentation Improvements"
  showcontent = true

  [[tool.towncrier.type]]
  directory = "other"
  name = "Other Changes"
  showcontent = true

  [[tool.towncrier.type]]
  directory = "author"
  name = "Contributors"
  showcontent = true

[tool.liccheck]
authorized_licenses = [
  "bsd",
  "new bsd",
  "simplified bsd",
  "apache",
  "apache 2.0",
  "apache software",
  "gnu lgpl",
  "gpl v2",
  "GNU General Public License v2 or later (GPLv2+)",
  "gpl v3",
  "lgpl with exceptions or zpl",
  "isc",
  "isc license (iscl)",
  "mit",
  "python software foundation",
  "zpl 2.1",
  "mpl-2.0",
  "lgpl",
]

[build-system]
requires = ["poetry>=1.0.0"]
build-backend = "poetry.masonry.api"